+++ app/handlers/newflow/educator_signup/complete_profile.rb
@@ -121,7 +121,7 @@ module Newflow
         end
 
         # Now we create the lead for the user... because we returned above if they did... again SheeridWebhook
-        Newflow::CreateSalesforceLead.perform_later(user_id: @user.id)
+        CreateSalesforceLead.perform_later(user_id: @user.id)
 
       end
