+++ app/routines/newflow/student_signup/activate_student.rb
@@ -17,7 +17,7 @@ module Newflow
 
         user.update!(state: User::ACTIVATED)
         if user.receive_newsletter?
-          Newflow::CreateSalesforceLead.perform_later(user_id: user.id)
+          CreateSalesforceLead.perform_later(user_id: user.id)
         end
         SecurityLog.create!(user: user, event_type: :user_became_activated)
       end
