+++ lib/tasks/accounts/create_leads_for_abandoned_user_signups.rake
@@ -10,7 +10,7 @@ namespace :accounts do
     users.each { |user|
       user.faculty_status = :incomplete_signup
       user.save!
-      Newflow::CreateSalesforceLead.perform_later(user_id: user.id)
+      CreateSalesforceLead.perform_later(user_id: user.id)
     }
   end
 end
