+++ spec/handlers/find_user_by_token_spec.rb
